Added settings command
